+++ pgmpy/parameter_estimator/__init__.py
@@ -1,7 +1,7 @@
-from .bayesian import DiscreteBayesianEstimator
-from .em import DiscreteEM
-from .gaussian import LinearGaussianMLE
-from .mle import DiscreteMLE
+from .discrete_bayesian import DiscreteBayesianEstimator
+from .discrete_em import DiscreteEM
+from .discrete_mle import DiscreteMLE
+from .linear_gaussian_mle import LinearGaussianMLE
 
 __all__ = [
     "DiscreteMLE",
